Bug: 804241 Reviewer: self Comment: Desktop application max file size needs to be a long, not int.

+++ ssf/main/src/org/kablink/teaming/remoting/rest/v1/util/ResourceUtil.java
@@ -371,7 +371,7 @@ public class ResourceUtil {
         desktopAppConfig.setAutoUpdateUrl(config.getFsaAutoUpdateUrl());
         desktopAppConfig.setEnabled(config.getFsaEnabled());
         desktopAppConfig.setSyncInterval(config.getFsaSynchInterval());
-        desktopAppConfig.setMaxFileSize(config.getFsaMaxFileSize() * 1024 * 1024);
+        desktopAppConfig.setMaxFileSize(((long)config.getFsaMaxFileSize()) * 1024 * 1024);
         desktopAppConfig.setAllowCachedPassword(config.getFsaAllowCachePwd());
         modelConfig.setDesktopAppConfig(desktopAppConfig);
